refactor: lua配下のファイル構成を再設計
- editor.lua → edit.lua（編集系）+ completion.lua（補完系）に分割
- keybind.lua を削除し、汎用keymapはinit.luaへ、mini.clueはui.luaへ移動
- mini.git / mini.diff を ui.lua から git.lua に分離
- mini.basics を nav.lua から edit.lua に移動
