+++ config/nvim/init.lua
@@ -17,9 +17,24 @@ end
 -- Set up 'mini.deps' (customize to your liking)
 require('mini.deps').setup({ path = { package = path_package } })
 
+-- keymaps
+vim.keymap.set('t', '<Esc><Esc>', [[<C-\><C-n>]], { noremap = true })
+
+-- 一括置換 :s<Space>
+vim.keymap.set('ca', 's', function()
+  if vim.fn.getcmdtype() == ':' and vim.fn.getcmdline() == 's' then
+    return '%s///g<Left><Left><Left>'
+  end
+  return 's'
+end, { expr = true })
+
+-- /検索のハイライトを消す
+vim.keymap.set('n', '<Esc><Esc>', '<cmd>nohlsearch<CR>', { silent = true })
+
 require('ui')
-require('editor')
+require('edit')
+require('completion')
+require('git')
 require('nav')
-require('keybind')
 require('lsp')
 require('oplog')
