--- config/nvim/lua/keybind.lua
@@ -1,76 +1,0 @@
--- キーバインド系
-
-vim.keymap.set('t', '<Esc><Esc>', [[<C-\><C-n>]], { noremap = true })
-
--- 一括置換 :s<Space>
-vim.keymap.set('ca', 's', function()
-  if vim.fn.getcmdtype() == ':' and vim.fn.getcmdline() == 's' then
-    return '%s///g<Left><Left><Left>'
-  end
-  return 's'
-end, { expr = true })
-
--- /検索のハイライトを消す
-vim.keymap.set('n', '<Esc><Esc>', '<cmd>nohlsearch<CR>', { silent = true })
-
--- clue (全体のキーバインドヒント)
-MiniDeps.later(function()
-  local function mode_nx(keys)
-    return { mode = 'n', keys = keys }, { mode = 'x', keys = keys }
-  end
-
-  local clue = require('mini.clue')
-  clue.setup({
-    triggers = {
-      -- Leader triggers
-      mode_nx('<leader>'),
-
-      -- Built-in completion
-      { mode = 'i', keys = '<c-x>' },
-
-      -- `g` key
-      mode_nx('g'),
-
-      -- Marks
-      mode_nx("'"),
-      mode_nx('`'),
-
-      -- Registers
-      mode_nx('"'),
-      { mode = 'i', keys = '<c-r>' },
-      { mode = 'c', keys = '<c-r>' },
-
-      -- Window commands
-      { mode = 'n', keys = '<c-w>' },
-
-      -- bracketed commands
-      { mode = 'n', keys = '[' },
-      { mode = 'n', keys = ']' },
-
-      -- `z` key
-      mode_nx('z'),
-
-      -- surround
-      mode_nx('s'),
-
-      -- text object
-      { mode = 'x', keys = 'i' },
-      { mode = 'x', keys = 'a' },
-      { mode = 'o', keys = 'i' },
-      { mode = 'o', keys = 'a' },
-
-      -- option toggle (mini.basics)
-      { mode = 'n', keys = 'm' },
-    },
-
-    clues = {
-      -- Enhance this by adding descriptions for <Leader> mapping groups
-      clue.gen_clues.builtin_completion(),
-      clue.gen_clues.g(),
-      clue.gen_clues.marks(),
-      clue.gen_clues.registers({ show_contents = true }),
-      clue.gen_clues.windows({ submode_resize = true, submode_move = true }),
-      clue.gen_clues.z(),
-    },
-  })
-end)
